Update magnetic field to match origin (U of MN flying field). Set magnetometer scaling to `1` and add method to set magnetic magnitude.
There are three ways to set magnetic field:
 1) Web lookup based on position: does not work as API has changed
 2) Specify the vector (e.g. look online and hard-code)
 3) Specify inclination/declination angle
Method 3 was incomplete because the magnitude was not being set.
This function was updated to use a specified magnitude.  Also
the magnetometer scaling was set to 1.  Previoulsy it was 2, which
doesn't seem to have any reason.

Note: the magnitude I'm assuming is in Gauss.  This seems to match
the expected units in other places of PX4 code but has not been
confirmed.